+++ example/lib/main.dart
@@ -25,13 +25,14 @@ import 'dart:core';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:http/http.dart' as http;
-
 import 'package:tru_sdk_flutter/tru_sdk_flutter.dart';
-
 import 'src/http/mock_client.dart';
+import 'package:crypto/crypto.dart';
+import '.env.dart';
 
 // Set up a local tunnel base url.
-final String baseURL = "YOUR_LOCAL_TUNNEL_URL";
+final String baseURL = "https://080a-2a00-23c7-8589-8d01-ed63-2488-2ec5-ed08.ngrok.io";
+
 
 void main() {
   runApp(PhoneCheckApp());
@@ -240,14 +241,33 @@ class _PhoneCheckAppState extends State<PhoneCheckHome> {
     );
   }
 
+  // Get Coverage Access Token
+
+ Future<TokenResponse>getCoverageAccessToken() async {
+    var signature = sha256.convert(utf8.encode(RTA_KEY));
+  final response = await http.get(
+     Uri.parse('$RTA_URL/coverage_access_token'),
+     headers: <String, String>{
+       'x-rta': signature.toString(),
+        },
+  );
+  if (response.statusCode == 200) {
+    return TokenResponse.fromJson(jsonDecode(response.body));
+  } else {
+    throw Exception('Failed to get coverage access token: No access token');
+  }
+  }
+
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<CheckStatus> executeFlow(String phoneNumber) async {
     print("[Reachability] - Start");
     var canMoveToNextStep = false;
+    var tokenResponse = await getCoverageAccessToken();
+    var token = tokenResponse.accessToken;
     TruSdkFlutter sdk = TruSdkFlutter();
     try {
-      Map reach = await sdk.openWithDataCellular(
-          "https://eu.api.tru.id/public/coverage/v0.1/device_ip", false);
+      Map reach = await sdk.openWithDataCellularAndAccessToken(
+          "https://eu.api.tru.id/coverage/v0.1/device_ip", token , true);
       print("isReachable = $reach");
       if (reach.containsKey("error")) {
         throw Exception(
@@ -360,6 +380,16 @@ Future<CheckStatus> exchangeCode(
   }
 }
 
+class TokenResponse {
+  final String accessToken;
+  TokenResponse({required this.accessToken});
+  factory TokenResponse.fromJson(Map<dynamic, dynamic>json) {
+    return TokenResponse(
+        accessToken: json['access_token']
+    );
+  }
+}
+
 class PhoneCheck {
   final String id;
   final String url;
@@ -387,6 +417,7 @@ class CheckStatus {
     );
   }
 }
+
 
 // Set up a mock HTTP client.
 final http.Client httpClient = MockClient();